+++ src/main/java/org/patternfly/core/Numbers.java
@@ -1,0 +1,36 @@
+/*
+ *  Copyright 2023 Red Hat
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+package org.patternfly.core;
+
+import java.math.BigDecimal;
+import java.math.RoundingMode;
+
+public class Numbers {
+
+    public static double percentage(double value, double max) {
+        return round((value * 100.0) / max);
+    }
+
+    public static double percentage(double value, double min, double max) {
+        return round(((value - min) * 100) / (max - min));
+    }
+
+    private static double round(double value) {
+        BigDecimal bd = BigDecimal.valueOf(value);
+        bd = bd.setScale(2, RoundingMode.HALF_UP);
+        return bd.doubleValue();
+    }
+}