+++ src/main/java/org/patternfly/style/Status.java
@@ -13,9 +13,7 @@
  *  See the License for the specific language governing permissions and
  *  limitations under the License.
  */
-package org.patternfly.core;
-
-import org.patternfly.style.Classes;
+package org.patternfly.style;
 
 public enum Status {